requirements.txt
aws-cdk.core>=1.134.0
aws-cdk.aws-ec2>=1.134.0
aws-cdk.aws-iam>=1.134.0
aws-cdk.assertions>=1.134.0
